Step 11 - Integrate mapper and repository

+++ data/src/main/java/com/coroutinelab/data/dto/emaildetails/Attachment.kt
@@ -13,5 +13,5 @@ data class Attachment(
     @SerialName("mimeType")
     val mimeType: String?,
     @SerialName("size")
-    val size: Int?
+    val size: Long?
 )
